Update routines.rb
Added login return value description from official API documentation.

Reworded BoundError message if macrobutton ID is out of range, for improved clarity.

Since create_alias relies on the the strip blueprints I've grouped them together in both initialize and login dependant on whether a type was passed as an argument variable or not.